// frontend/vue-project/src/api/models/virtualFileVO.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FileItemType } from './fileItemType';

/**
 * 用于前端展示的视图对象
 */
export interface VirtualFileVO {
  /** 文件/文件夹ID */
  id?: number;
  /** 文件/文件夹名称 */
  name?: string;
  /** 文件大小（字节） */
  size?: number;
  /** 最后更新时间 */
  updateTime?: string;
  type?: FileItemType;
  /** 父目录ID */
  parentId?: number;
}